unifying macros for defining problems and dimensions, and added warnings for dimension mismatches in the defined problems.
